Allow ENVied to be fully configurable via ENV
ENV['ENVIED_GROUPS']:
This ENV-variable is taken into account when calling `ENVied.require`
*without arguments*.

ENV['ENVIED_ENABLE_DEFAULTS']:
this ENV-variable is taken into account for enabling defaults.
As soon as you call `enable_defaults!` in the Envfile (with whatever
arguments), you overwrite this value.

So in both cases the values of these variables serve as a default.