Implement our own coercion functionality
This also includes the error related code and matches the extensive tests in coercible to make sure we keep the same functionality and be able to safely remove the coercible dependency.

For the string tests, include a failure case for if the value does not respond to the `to_str` method.

We implement our own coercer for integer so we don't need the full test cases from coercible. These should give us plenty of coverage. Now these include an instance type check and invalid data case too.

For the float type, this functionality we were getting from coercible but we have a different implementation in place. All the same test cases pass including the failure ones.