Replace coercible dependency with local solution
We now include directly just the functionality we use from the coercible gem so all references and the dependency have been removed safely.

We just relied on the "string->other" coercion types which are rather simple for us to implement. We have also brought over all the test cases for each from coercible so we have the same coverage on our end.

Note our float implementation is different from coercible as we just use the Float() constructor to coerce a string to a float. These passed all our existing tests so its not considered a breaking change.

Simplify the coerce method to now just take 2 arguments: the value and type.