Update generated gem files with current changes
These updates the generated files when creating a new gem with what's current including setup/console scripts. This updates the .gitignore with the defaults and adding the Gemfile.lock. For some reason it included the root bin directory despite being in git history, we want that though so we can remove it.