Update tests and incoercible error message format
The error message already starts with that the following variables are not coercible so we can just list the name, value and type. The test case now accounts for multiple incoercible variables.

This reorganizes helper methods to isolate them where used.

Instead of using stub_const in test cases, updated to use the and_ENV helper.

Made a note on changing order so the and_ENV is called first before configuring.

Do not use the `to be` form with no explicit value.

Now set the ENV first before configuration as that makes most sense, otherwise we have to have test cases using the stub_const directly because it's need to be set first. This is more flexible and clear as to how it is expected to work.

Updated all test ENV variables to now be in uppercase to reflect how they should be setup.

Updated tests that check for errors to now verify the full error message.