Mark attr_readers as private
We want to use these within the class but not expose them through the proxy. The `config` is already exposed as a reader in the ENVied class so it won't ever get to this point but we don't want `coercer` exposed either.

To set `groups` as private we need to revisit the tests in the envied_spec file since its testing against that directly but it shouldn't be necessary. We should have coverage on groups functionality already if not we can expand it.