TEMP - set a warning
The conditional ignores procs. What to do?

Only show warning when retrieving the default value through ENVied class/proxy?

Maybe show in both cases but for the variable class, only show in the case the value is a non-proc.

Updated so if default is a proc, when value is retrieved, we show a warning if its a non-string value.